+++ src/main/java/spring/fitlinkbe/application/auth/AuthFacade.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import spring.fitlinkbe.domain.auth.AuthService;
 import spring.fitlinkbe.domain.auth.command.AuthCommand;
-import spring.fitlinkbe.domain.common.PersonalDetailService;
 import spring.fitlinkbe.domain.common.model.PersonalDetail;
 import spring.fitlinkbe.domain.common.model.Token;
 import spring.fitlinkbe.domain.member.Member;
@@ -18,7 +17,6 @@ import spring.fitlinkbe.support.security.AuthTokenProvider;
 public class AuthFacade {
 
     private final MemberService memberService;
-    private final PersonalDetailService personalDetailService;
     private final AuthService authService;
     private final AuthTokenProvider authTokenProvider;
 
@@ -26,7 +24,7 @@ public class AuthFacade {
         Member savedMember = memberService.saveMember(command.toMember());
 
         // personalDetail 업데이트
-        PersonalDetail personalDetail = personalDetailService.registerMember(personalDetailId, command, savedMember);
+        PersonalDetail personalDetail = memberService.registerMember(personalDetailId, command, savedMember);
 
         // 토큰 생성 또는 업데이트
         String accessToken = authTokenProvider.createAccessToken(personalDetail.getStatus(), personalDetailId);
